refactor: BaseFrame 추가. NormalFrame, FinalFrame 중복제거
