+++ src/main/java/bowling/domain/FinalFrame.java
@@ -1,33 +1,9 @@
 package bowling.domain;
 
-import java.util.List;
-import java.util.stream.Collectors;
+public class FinalFrame extends BaseFrame {
 
-public class FinalFrame implements Frame {
-
-    private final int frameNo;
-    private final Frame prev;
-    private PinMarks pinMarks;
-
-    FinalFrame(int frameNo, Frame prev) {
-        this.prev = prev;
-        this.frameNo = frameNo;
-        this.pinMarks = new FinalFramePinMarks();
-    }
-
-    @Override
-    public int getFrameNo() {
-        return frameNo;
-    }
-
-    @Override
-    public void mark(int countOfFallDownPins) {
-        pinMarks.mark(PinMark.pin(countOfFallDownPins));
-    }
-
-    @Override
-    public long getCountOfMarks() {
-        return pinMarks.getCountOfMarks();
+    FinalFrame(int frameNo) {
+        super(frameNo, new FinalFramePinMarks());
     }
 
     @Override
@@ -44,41 +20,8 @@ public class FinalFrame implements Frame {
     }
 
     @Override
-    public boolean isOpen() {
-        return false;
-    }
-
-    @Override
-    public boolean isEnd() {
-        return pinMarks.isAllMarked();
-    }
-
-    @Override
     public Frame createNext() {
         return null;
     }
 
-    @Override
-    public Frame next() {
-        return null;
-    }
-
-    @Override
-    public FrameScore getScore() {
-
-        return FrameScoreCalculatorFactory.create(isFinal(), getStatus()).calculate(this);
-    }
-
-    @Override
-    public FrameInfo toFrameInfo() {
-        return FrameInfo.of(getFrameNo(), pinMarks.toSigns(), getScore());
-    }
-
-    @Override
-    public List<Integer> getCountListOfFallDownPins() {
-        return pinMarks.toList()
-                .stream()
-                .map(pinMark -> pinMark.getCountOfFallDownPins())
-                .collect(Collectors.toList());
-    }
 }
